requirements.txt
d2to1>=0.2.10,<0.3
pbr>=0.5.10,<0.6
amqplib>=0.6.1
anyjson>=0.2.4
argparse
babel>=0.9.6
eventlet>=0.9.12
greenlet>=0.3.1
iso8601>=0.1.4
kombu>=1.0.4
lockfile>=0.8
lxml>=2.3
oslo.config>=1.1.0
paramiko>=1.8.0
paste
pecan>=0.2.0
python-glanceclient>=0.5.0,<2
python-keystoneclient>=0.2,<0.3
python-swiftclient>=1.2,<2
routes>=1.12.3
six
sqlalchemy>=0.7,<=0.7.99
sqlalchemy-migrate>=0.7
stevedore>=0.9
suds>=0.4
webob>=1.0.8
wsme>=0.5b2
wsgiref>=0.1.2
